Move TemperatureController from extruder to separate files
